Publish releasenotes
Enhance releasenote handling and add a post job that publishes the
release notes.

Note that the post jobs will be run on all branches but we need
to publish the content from master since we generate all content
on the master branch.

Add job to post queue for publishing release-notes.

Change-Id: I204e669d532d59bbf6aec0d8d4da4ab183deaf59