Add networking-bgpvpn jobs for bagpipe driver
This change defines a jenkins job template to test the installation
of the networking-bgpvpn bagpipe driver (the existing job was only
testing a dummy driver).

This template is then used to define jobs to run the checks both against
the master branch of other Openstack projects (in our case devstack and
neutron import) and against the stable/liberty branch of other Openstack
projects. The motivation for doing so was explained in detail for
I049a521b4e948a042c9b49eaecd9399cf84353d1.

Change-Id: I7273a5e486e7023a59b95a325db2e6400c209a48