Add rake-dsvm gate job-template
To continue work on integration of Fuel-* projects in
OpenStack infrastructure.

We have two ruby component in Fuel: fuel-astute and
fuel-nailgun-agent. We use rspec for unit tests.
We can not use gate-{name}-rake which use trusty and ruby 1.9
because our gem deps require ruby 2.0
We can use puppet-unit jobs on devstack with centos7
but it looks useless to install puppet before running rspec.

So we offer to make the ruby unit test job on centos7.

This patch creates a basic ruby unit test job
which setups bundle and runs rake spec on dsvm

Change-Id: I13991cd5c145bf968069f4b7ac706c9ffecb953b