Shorten the kolla job names
Due to the length of the job name and the tox target we run into an
uncommon limitation; the virtualenv that tox launches is nested in a
path that is too long. This leads to an error on our longest named
job which prevents tox from running properly.

This limitation is the limit for the line length of the first line
in a shell script. In this case <virtualenv>/bin/pip has a path that
exceeds that limit by a few characters. This results in a very
confusing "OSError: [Errno 2] No such file or directory" error.

See `man execve` for more info. A quote from that manpage:
'A maximum line length of 127 characters is allowed for the first line
in a #! executable shell script.'

To reproduce this locally you can just create a very long path and
then run tox in it. Example: http://paste.openstack.org/show/478559/

Change-Id: I51f8f67d9ab139f1d539fce05ce3cc4766aeedff
Depends-On: I43fba2a5ff1890d699045496c9eaee5e849f3e75