Fix grab console log partial gets
The grab console log script started by cleaning out any existing console
log files with echo. Unfortunately echo appends a newline by default
which meant we ended up with a 1 byte file instead of a 0 byte file.
This resulted in grabbing the console log from byte offset 1 instead of
0 and jenkins is weird and prepended base64 content at the beginning of
the log file in this case.

Fix this by using echo -n to clear the file which results in a 0 byte
file.

Change-Id: I8bd7cbd0c19cd44067839b34e317248daad9f578