Add os-performance-tools
This code has been written to push performance counters from devstack
runs into statsd. It is intended to work as a plugin for subunit2sql,
but could very easily be adapted to receive code from other streams as
well.

This project should not be added until the corresponding spec is merged.
That spec is marked as a Depends-On below.

Depends-On: I3ad971cfa33789ade733e8fb92ff51d487078638
Change-Id: I2019afa4ef4b6247b9dd2139f4df98196370f4fe
Implements: counter-inspection