Creation of the tools-contrib repo.
This review creates the tools-contrib directory where it will be a place
to contribute any tool, script, or the like to the OSOps project. The
rename of the tools-generic as a Curated place, is where we plan on
putting coding standards and testing around the tools.

These both were discussed in Tokyo, and decided that we needed a place
for each type.

Change-Id: If8c848fcac4e1a86e0d381883b87cf9dbfdf8b36