swap 3 tempest jobs for 1 layer4 jumbo job
The tempest layer4 job runs all the tests from the heat / sahara /
trove jobs, but in one config. It's run successfully on all the
experimental runs so far.

This will provide us with the same test coverage to ensure these tests
can work on the heat / sahara / trove jobs, but will separate them
from the integrated gate because tempest will now not share a job, by
name, with heat / sahara / trove. It should create faster merge for
all projects because of this with no loss of test coverage.

Change-Id: I34d6427d294cdf82ddf1d4709d3e1630551959f9