Add neutron -plus job template for advanced features
This job should eventually go into neutron gate as voting. The job is in
neutron experimental queue for now. Once we validate it works fine,
we'll enable it as non-voting in the project gate, and finally enable
votes for the variant.

We could not replace the existing job that does not have a hook since
this job belongs to the integrated gate, and we are not allowed to
depend on hooks there.

Change-Id: Id92723cc88428d370f5e9ac5d79a7a08f389a2d0
Depends-On: I9f73b938539586022aeb22e3a32877816e1dcaba