Fix two errors in setup_training_guides
setup_training_guides was written assuming that the project would be
passed as the first argument, but it wasn't called like that. Change
the function to match others to hardcode the project, and look for the
version as the first argument.

Correct the path to the locale directory so that the Zanata client can
read and write translation files correctly.

Change-Id: Iacfd2e831d18af5870fc70a2a54e569d0924474e