Fix job deffinition for ceph job
Recently, I4f70a3785ea4aeb38c3ebd8fd6a74ed3d035fa1a landed and fixed the
definition for the ceph-src job. This fix copies the same fix from that
job into the non-src ceph job as it's been failing in stable branches
with the same footprint.

The change removes the TEMPEST_FULL definition as the original fix did
and it changes the regular expression to align both jobs.

Change-Id: If289bfe3a5e4958411c8fae41a73971433c00d8b