Remove non-voting requirements check job from cisco plugin
The requirements check job template is meant to be voting on
both check and gate should a repo choose to use the template.

Exceptions to this usage cause confusion and cargo culting.

No repo is required to run the requirements check job template
but should they choose to they are expected to use it as intended.

The requirements check job template is needed for a project to be
added to global requirements.

Change-Id: I5c1938a8069f8848069f5d30ce82aa010a84f569