Add check-requirements for cliff
cliff had made requirements check in
I40b2cf4030df66c851d7d11da4f81f645d110e76 and it was not enabled again.

Remove non-voting and make it voting in the gate as well - with using
the check-requirements template.

Note that cliff is in the global requirments sync, thus it should have a
gate job as explained in
http://docs.openstack.org/developer/requirements/#enforcement-in-projects

Change-Id: I4f22c58e469d42b64694123b13c64fe620bae1eb