modify openstackclient functional tests to use neutron dsvm
OSC is finally getting ramped up to create neutron support, to do
so effectively, we want to create functional tests, so switch our
functional test job to use neutron.

I opted to directly change the job and avoid adding a second experimental
job. The chances of this breaking OSC is very low, and affects only
functional test code. If the gate does break we can simply skip the
tests that are affected so the gate is passing again.

Change-Id: I88972f535d3b10326c3e848d569d2e93564898f0