Add ansible-role-(diskimage-builder|jenkins-job-builder)
Here we are just adding ansible-role-diskimage-builder and
ansible-role-jenkins-job-builder into the OpenStack gerrit workflow.

After talking with both openstack-infra and third party CI teams it
was decided to deal with which team this would be part of a later
date.

Additionally, we have not created a launchpad project for this module
or pypi (since we won't release there).

Gerritbot was also not configured at this time, since we don't know
which team we are going to be part of.

Change-Id: I18397f709059e3c4455f3fe9df2be08a6993fe7f
Signed-off-by: Paul Belanger <pabelanger@redhat.com>