Remove py26 support from novaclient
as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: Id93b7af9cc3ed548b78e9b08528e18e9c0383c7a
Closes-Bug: 1519510