Remove py26 support from python-neutronclient
as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I4a2f860e8138bddb032418ee555e07a6fa0a3f3e
Closes-Bug: 1519510