Remove py26 support from cinderclient
as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I6f9f7dcb4e454f4d1e12943a1f4a1ab3ced0eba3
Closes-Bug: 1519510