Remove py26 support from designateclient
as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I479081c88b53175def2a5ea06ee1d3435ed6444e
Closes-Bug: 1519510