Remove py26 support from fuelclient
as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I3ea60260b467710a3e287d88e026f72060f63a1d
Closes-Bug: 1519510