Remove py26 support from ganttclient
as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I3ecb1a28a0dcde0c7a95d4aa6efd2dc058fa2dbd
Closes-Bug: 1519510