Remove py26 support from glanceclient
as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I53de6c93d8fd3a5e7afd57bbeb175a91becd2e3f
Closes-Bug: 1519510