Remove py26 support from heatclient
as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I745d671fa5092fb5b796b18674e21dab1991643c
Closes-Bug: 1519510