Remove py26 support from kiteclient
as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: Ie6bd472cb3d2947bb5b9e1fec982a65cfced624e
Closes-Bug: 1519510