Remove py26 support from muranoclient
as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: Ibea6301c8c1983b6a2a1e86c15de83b6684ed3d6
Closes-Bug: 1519510