Remove py26 support from saharaclient
as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I086b29d3d2a165d9f656564640103c097b83e153
Closes-Bug: 1519510