Remove py26 support from swiftclient
as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I37116731db11449d0c374a6a83a3a43789a19d5f
Closes-Bug: 1519510