Remove py26 support from troveclient
as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: Idf794853a05e46d6f24132ba6ac1e508a035d616
Closes-Bug: 1519510