Remove py26 support from tuskarclient
as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I7ad35b9007730237494c4ae38e41f851e16b295d
Closes-Bug: 1519510