Use WARNING style for JJB compare XML job failure
First word of the job status message to Gerrit will be used to
construct the style to be applied. Using 'WARNING' will activate the
result_WARNING style to be applied from the Gerrit site style sheet.
This should in turn colour the word 'WARNING' differently to ensure it
is easily detectable when the output differs between runs.

Change-Id: Ie7c02c897c48e434f59f530f46d00e3291ea2ef6
Depends-On: I1a9bea0c0830f9ca15f8369632ff2fdbeadada56