Add sdist check to pep8 build script
Inability to create an sdist tarball is a fundamental problem, but
because of how we run tox in unittest builders, we often do not catch
this anymore. The potential harm is that a change can land to a project
that breaks the ability to create an sdist or run a direct setup.py
install or pip install . - which will in turn break the requirements
gate which tries to do those things.

Change-Id: Ibc36edb06cf072b9f0cf664fae3c3c2273ff71e6