add 'tempest' in puppet-openstack-integration jobs
To have the puppet integration jobs part of Health [1] project, we want
logstash about to read subunit-files.
So we need to rename the jobs to match this job filter:
job-filter: '.*(tempest|grenade)-dsvm.*'

That way, we will be able to see the subunit results on [1].
Puppet group already did the necessary to prepare the subunit files, we
just miss this patch to actually rename the jobs.

[1] http://status.openstack.org/openstack-health/#/

Change-Id: Ia41489a913d2adfc00b5f0d9177d1e810e95eec8