Add separate non-voting job to check doc links for JJB
Add's a non-voting job to catch both invalid links being added, and
valid links going stale over time in the JJB documentation built using
sphinx.

Needs to be non-voting to avoid requiring the external resources are
always reachable for patches to pass. In the future sphinx should
contain a change[1] to ensure that warnings in the linkcheck builder
will become errors when the warniserror option is enabled, which will
make such problems with external resources fatal.

[1] https://github.com/sphinx-doc/sphinx/issues/1815

Change-Id: I7526b737523249b07defc03f19998c8f615bdddb
Depends-On: I80446abe96824a444a04cc526df6cc38632a4607