Adding trove-dashboard repo and jobs
The Trove content from Horizon is moving to a new repository to
be developed as a plugin to Horizon. This change sets up the
repository and the gate jobs for the new repository so the code
can be moved.

Used git filter-branch to maintain history and created temporary
repository with the results:
    https://github.com/dklyle/trove-dashboard
On top of the history and moved files, considerable moving of files
and additional pieces were added to complete the plugin.

https://pypi.python.org/pypi/trove-dashboard has been created with
the appropriate permissions for openstackci.

Needed-By: Ief3829b45987dfb4078c78428f5f699912542868
Change-Id: I158eac980f767a4b8cd43294e8cc215fbb1586f3