Test that job regexes in zuul/layout match
Enhance our zuul/layout.yaml checks to check that the names in the jobs
section actually match an existing job. A mismatch implies that the
pattern does not work as intented.

Add note to run-layout.sh that the job files is used elsewhere as well,
so that it does not get deleted by accident.

Change-Id: I5c4cd97829acf90bb8f72dde32e5fe8063933cdb