fix stable/liberty telemetry integration job
devstack-gate sets up the environment for gate jobs but it
configures branches at a global level rather than a per project
level. this means projects which are released independently can't
be tested with other projects.

this patch skips the devstack-gate configuration of gnocchi and
lets devstack configure it so enable_plugin functionality is properly
handled.

Change-Id: I924f99ffb3344b2d501c53dbbc8403f419dfdee7
Closes-Bug: #1526530