Disable reno on liberty for keystoneauth/keystonemiddleware
As Doug said in
http://lists.openstack.org/pipermail/openstack-dev/2015-November/080694.html,
reno setup for libraries is optional for stable/liberty.

python-keystoneclient, keystoneauth, and keystonemiddleware currently
do not handle release notes on stable/liberty, so do not run reno there.

Keep the post job to run on any change. It always checks out master, so
there's no harm running it when liberty changes merge.

Change-Id: Idbd047746dbead5af095d5d69544593f285c3bce