Check for dead links in App Catalog
This commit includes a job meant to be run periodically which checks
all links to assets listed in the Community App Catalog assets.yaml
file. For any dead links, an additional attribute is added to the YAML
file, and a review is submitted with the updated YAML which will
prevent the link from being displayed until it's updated with a
subsequent review.

Change-Id: I991155edca8ed92431f7a5deacf2d3938cf70ada