Skip OpenStack-Ansible releasenotes job for <Liberty
It appears that releasenotes build jobs are executing on all
branches. This prevents the job from executing on our 'juno'
and 'kilo' branches as neither of them are instrumented for
releasenotes builds.

The existing filter work on stable/kilo and stable/liberty
but ansible uses kilo and liberty, thus needs its own filters.

Change-Id: I42dbd4e5a21ea739006f017b94965cfb45942ac3