Remove precise/trusty label mapping in Zuul config
Now that there are no jobs declaring both .*-precise and .*-trusty
nodes at the same time, it is safe to remove the associated mapping
in Zuul and let it trust that any worker registering the job name
is capable of running it sanely.

This should not be approved until manual jenkins-jobs updates have
been completed and confirmed successful on all Jenkins masters.

Change-Id: I3b40021ffbf89842ed3a53412dfffa119f311797