Unify usages of checking for existing changes (1/2)
Create new function setup_commit_message in common.sh that sets the
commit message after checking for its existance. This check is
currently done in four places, the code has been copied around for this.
Let's unify this so that we only have a single piece of this code block.

For this change, use  the new function only for translation update.
The next change will update all other places. We're doing this in steps
in case of problems.

Change-Id: I80c7f81b0b7f225fbf1e1564239f40b5df8537d9