Minimal translation setup for murano-dashboard
Murano-dashboard is a dashboard for horizon, therefore shares the same
message structure as horizon does, i.e. django and djangojs .pot files.
Therefore this commit alters jenkins scripts, that update
translations, to include murano-dashboard where required.
However murano-dashboard (unlike horizon) does not use a shell file to
run it's tests, instead relying on tox.

This change does not actually run the jobs, we need to have a repo
for testing that the new consolidated setup works, so getting this in
helps with some manual testing.

Targets blueprint: murano-i18n

Change-Id: I7e8a620c12bcf47a908a4011976d1b148ee3e0dd