Merge openstack-ansible skip-if to resolve conflict
The current rules have a conflicting match, resulting in the skip-if
for the docs jobs never taking effect.

This patch merges both skip-if rules so that if any patches contain
docs and/or release notes and nothing else, then only those jobs will
run.

The purpose of this skip is to save the unnecessary consumption of
nodepool nodes.

Change-Id: Idc3c9d06c503d3e84395e193345023c9aeab21c6