stop proposing constraints updates to stable branches
We want to manage constraints in stable branches by hand, instead of
having the bot update them for us. This will help us stay closer to the
versions we released with, which is what we expect folks to be using in
production. When we have the stable library spec [1] implemented, we may
restore these updates for upper-constraints.txt because we will have the
separate release-constraints.txt list.

[1] https://review.openstack.org/#/c/226157/

Change-Id: I97f0cf78bf2c980628abe0f18aa9f5e96af70ae7