Test gate-grenade-dsvm-partial-ncpu on requirements repo
We currently only have partial grenade testing on mitaka
changes to the requirements repo, but we should also have
this testing on stable/liberty changes, the problem is in the
job definitions between liberty and mitaka that do that testing.

This simply adds gate-grenade-dsvm-partial-ncpu to the check
and gate queues for the requirements repo changes. Since
the job filters are such that they should be mutually exclusive
already, this should not step on mitaka.

Change-Id: I8b4e46ed001ce6af504f8462a3049f79f17684df