Keystone-only uwsgi job
This adds a non-voting gate job to keystone master that has keystone
running under uwsgi rather than a) mod_wsgi in Apache httpd (which
all the other projects use by default), and b) eventlet (which we
have a keystone gate job for already).

We're planning to eventually remove support for eventlet but we
should test running keystone under different wsgi containers to
make sure it's compatible.

 * non-voting so we can see that it works.
 * master-only since stable branches don't have the devstack
   change.

Change-Id: I732792ebd00a501ef91be64b719ed6ce1dc1d24f