Grafana: Update OVH graph with axis labels
Add axis labels and units where appropriate.

Change the launch attempts graphs to summarize to 1m rather than
1h since grafana lets us zoom in.  1m is the lowest native unit
of time that will always show whole numbers for this metric (whose
lowest non-zero value is 1 event / 10 seconds).

Change the test nodes graph to stacked to match the way we normally
draw this graph, but change the tooltip to 'individual' so that
when hovering, individual values for the different states are
displayed, rather than cumulative (which does not make sense for
this application).

Also change the tooltip for the node graphs on the zuul dashboard
in the same manner.

Change-Id: I500aa486362476cff76a3d254093723f27021bed
Depends-On: Ie542dc4d0e151a00e84cc970c2cfa8c02377d7bf