Allow ironic jobs to run with tempest plugin
Ironic is migrating to a tempest plugin. This change sets
DEVSTACK_GATE_TEMPEST_ALL_PLUGINS=1 so that the plugin is loaded, and
also changes DEVSTACK_GATE_TEMPEST_REGEX to include tests with 'ironic'
in the name, as the plugin doesn't have 'baremetal' in all test names.

This patch retains the current 'baremetal' regex for the stable/kilo and
stable/liberty branches. That support can be removed in the future.

Depends-On: Ic52806987dae9f9df561ebd662f12c3445d0e2af
Change-Id: I535d7f8170706b3af19ee8c060493c8246e2ce31