remove gate-tempest-dsvm-ceilometer-mysql-neutron-full from tempest gate queue
This job is coupling the ceilometer cluster of projects to the
integrated gate through transitive dependencies. Which means a lot
more gate queue backup for projects that largely don't impact each
other.

This job is left on the Tempest check queue to ensure we don't
completely break Telemetry projects, but removed from the gate.

Change-Id: I4495b2b23d58a0bda35f7c663f93804913368a9b