Run the ofagent ci job on the periodic-neutron queue
As the ofagnet depends on the neutron, the ofagent is affected by a change
for the neutron and some changes may break the ofagent.
So we want to run periodically ofagent job with the latest neutron repo
to verify that the ofagent isn't broken them.

As the ofagent ci job is stable, we put it to the periodic-neutron queue.

Change-Id: I0c6d371edd77e42c9cde6bf79bbd07164733e50f