disable voting for tempest-dsvm-keystone-eventlet-full
This job has been failing too often lately, each time with a timeout
related failures.

For example: http://logs.openstack.org/68/264068/6/check/gate-tempest-dsvm-keystone-eventlet-full/947b46b/console.html#_2016-01-17_21_42_48_600

Jobs that are passing are taking around 1h 44m, but the ones that
fail seem to be around 1h 58m.

This is impacting keystone development as approved code is
failing gate and check.

Temporarily disable voting of this job until we can identity
the cause of the slowdown and ensure it passes consistently.

Change-Id: I3f0d3718a62d7ebf093c4da80d983692b96b7dfe