puppet: initiate testing for puppet-openstack-spec-helper
Puppet OpenStack group has a library [1] that is used by our modules and
contain some common ruby code.

This library really help us to:
* be consistent in how we test our modules
* avoid duplicated code

So far, we've put Rakefile and beaker helper in the gem, and we're in
the process to use it in our modules.
It's dropping lot of code! And so far we like it.

The problem is it's very easy to break the library because it has no
testing yet. If you break the library, you break all Puppet CI.

This patch aims to initiate it:
* create a job: gate-puppet-openstack-spec-helper-unit-dsvm-centos7
* run this job only on puppet-openstack_spec_helper project
* set the job as non-voting
* the job will just run ./run_unit_tests.sh so we will handle the rest
  in openstack/puppet-openstack_spec_helper repository.
* this patch only focus on unit tests. Further patches will come for
  beaker and eventually more bits we want to test.

[1] https://git.openstack.org/cgit/openstack/puppet-openstack_spec_helper/

Change-Id: I9c7d9efae0907e4cc4c14d76dcf44ef32f78a785