Run ovs-native jobs on the periodic-neutron queue
ovs driver with native of_interface isn't tested on tempest job
by default. We want to run periodically ovs-native jobs to verify
that the ovs driver with native of_interface works well with
the latest neutron codes.

Currently ovs-native jobs are run on the experimental queue and
it seems to be stable. So we can run them on the periodic-neutron
queue.

Change-Id: I60df32e200200b42a3b54167538dc9315592fa52