Add a tinyIPA non-voting test to Ironic
TinyIPA is a small tinycorelinux based IPA image that Ironic is
considering using instead of the coreos image, this patch adds a
non-voting test to the Ironic check pipeline to start the ball rolling.
This change depends on tinyipa merging into the ironic-python-agent
project, as well as an update to the ironic devstack plugin.

Change-Id: Ibfcd715538275ee0278fd392dad80f7341890951
Depends-On: I79fbec75de19b3d6feb19f87a3a0f662eae42bf3
Depends-On: If45c36cb8dc267dc80d1a324e8db3286ba6f7ee8