Make TripleO puppet check jobs voting
These are all green now, modulo the regression referenced in the
dependency below, so we can switch these jobs voting to avoid folks
ignoring the results.

Note that these jobs are also added to the gate queue, the unit
and style tests should be fast enough to run on gating, but we
don't want to re-run check-tripleo-jobs because these functional
tests take a long time to run.

Depends-On: I3cdcca6364476b324249a778d5e75a8befa26dff
Change-Id: Ida4aebcf5bebb8f83e6d446ef8db4dea56cc855f