increase large-ops job timeout
The large-ops overall jenkins job timeout is 60 minutes. However,
since the move to upper-constraints, and the lack of cached wheels,
we're spending 13 minutes installing python packages. This is very
often tripping use past the 60 minute timeout very barely. Bump this
up for more headroom so that we only fail if tempest actually fails,
not because pip is slow.

Change-Id: I49365b972e7095e3a65d9483181814c42de5df8b