Make the rally job experimental
This job has not progressed significantly in the time it's existed
and remains relatively uninteresting, the results vary quite a bit
depending on the cloud/node it is running on.

The move to experimental means that it can still be triggered
manually if desired, but removing this will free resources for
other interesting check/gate jobs.

Change-Id: I30fa34cf7685bc3a81ecfd796816ab945af17b1b