Don't try to filter release jobs by branch
Remove some branch filters which were intended to
compliment post and release pipeline jobs removed in
I23fd9e0af81c5c12da2c8daca214ecca8ee173d2 already. For
release pipeline jobs this was particularly incorrect
anyway, as the refname will be the tag and Zuul does not
know to map this back to a branch to filter in the first
place.

Change-Id: Ib456870627aea895d636da7898a5fac5006da544