Remove some jobs from gantt
The gant repos are inactive, remove some jobs from them. Even if they
are not run, changes to e.g. integrated gate template might need
changes for gantt. Remove translation jobs (gantt is not translated),
as well as integrated-gate (client is not part of it anymore, this only
confuses), grenade, large-ops.

Change-Id: I69a7bb7ebe0bf83959033810e1b3795ffe47f953