copy_puppet_logs: use SUDO to rename config/log files
In a previous commit [1], we stopped changing the owner to manually set
file mode bits. The problem is that 644 is not enough for files, because
they need to be renamed so write permissions are needed.

This patch aims to make sure files can be renamed by any user, so we
use SUDO prefix for each 'mv' command.

[1] I1c93a841b82514fb146d827332eecea98d9dbd3b

Change-Id: I2ea4b090c99af33dd0095cbe664cd4187b114b54