zuul/layout: optimize puppet jobs
This patch aims to optimize the consumtion of OpenStack Infra for Puppet
OpenStack jobs.

It will avoid to run some jobs if we're just patching *.md files or
releasenotes managed by reno.

Here are the job we want to skip:
* gate-puppet-.*-puppet-syntax-.*-dsvm-.*
* gate-puppet-.*-puppet-unit-.*-dsvm-.*
* gate-puppet-.*-puppet-beaker-rspec-dsvm-.*
* gate-puppet-openstack-integration-scenario00(1|2)-tempest-dsvm-.*

Which are syntax, unit & functional testing jobs.

It will save a lot of resources when we just want to update releasenotes
or doc.

Change-Id: I300abb47259a6fa9acdcfbc3e2eb15f9c3fea69b