Use new translation setup for all python projects
Use the new translation setup for all python projects. This now uses
it for most python projects and those will just work without any change
with the new setup since for them modulename and reponame are the same.
For example for nova, both are "nova".

In addition, enable new setup for vmware-nsx and ironic-inspector - the
projects are ready now.

Note this change temporarily disables python-novaclient and
python-neutronclient - instead of making the script complicated, we wait
until all python projects are ready and then reenable all.

This also removes the old propose_python function that is not needed
anymore and renames propose_python_new to propose_python.

Remove also some outdated comments.

Story: 2000452
Change-Id: I5b89f7b168a862773282dc0400763b1a94ea9e10