Re-Enable django_openstack_auth/designate-dashboard translations
Also, remove liberty jobs for these so that we do not need to backport
or handle them extra.

Backporting would break packaging by third parties, therefore not an
option.

Designate-dashboard setup is broken in liberty anyhow, so no need to
handle it extra. django_openstack_auth is good translated, and we're at
the end of translation.

Besides re-enabling the translations using the new setup, we can now
remove the old function setup_django_openstack_auth since it's not used
anymore.

Story: 2000452
Change-Id: I8b72a148d0f0e89d61807eee50a772e1fbcecaa2