Add openstack/broadview-lib project
broadview-lib provides a Python interface to BroadView, which
provides physical network switch (underlay) metrics. broadview-lib
supports the configuration of BroadView agents, and supports parsing of
monitoring messages published by agents to a collector or application
that receives them. Received messages are converted to an object model
that can be used to get at metrics data.

This project is the first of an overall contribution that will supply
metrics data to monitoring projects like Monasca.

Change-Id: I329926e69202a2e1f37e5e0bc9e5891db5589484