Atomically replace the wheel index file
Even though users of the read-only volume will see atomic updates
of the mirror, users of the read-write would be able to see the
index file as it was built.  Additionally, since OpenAFS syncs on
close, each >> concatenation would result in a sync meaning that
the process would take longer than needed.

Instead, do a traditional atomic mv operation to put the index
file into place.

Change-Id: I8d5d49c6158f6d6ae7f0bef72a2f64500cb425a5