Experimental periodic jobs for testing against oslo.* from master
Starting with Nova (from master), we'd like to test py27 and py34
tox targets against oslo.* from master branch on a daily basis.
Once we get this stabilized we can add other projects as well.
Note that we only test master of a project against oslo.* master
in the two jobs defined in this review.

Change-Id: I82ef7a5d5a3c1076efbd4306b7099cd28960cd90