Add non-voting CentOS check to OpenStack-Ansible roles
As OpenStack-Ansible is preparing for broader OS support, the ability
to do checks against CentOS provides a good validator to confirm that
the submitted patches are working.

For now this check is being added as a non-voting check. Once each
role has been confirmed to be ready this will be converted to a
voting check.

Change-Id: I8c6aa9c71996d3ccf5d8d28ff9c9af1cad9e20f2