Watcher multinode Tempest job fixup
Following https://review.openstack.org/#/c/274068/, I have an issue
with the "gate-watcher-dsvm-multinode-nv" job which is currently
failing when trying to deploy the subnodes.
The declaration of ENABLED_SERVICES as it is currently the case seem
to be the cause of it as it installs the list of services on both
the controller node and its subnodes.

This patchset hence removes the declaration of these services.

Change-Id: I92a6d9a2b03dfb1560031f2dd790c53863519187
Related-Bug: #1538606