Added *-npm-publish-draft job
This job performs the same work as the js-draft series,
with the added benefit that it does not require knowledge
of the underlying project hierarchy, nor custom hooks into
grunt to generate a release package. It calls the NPM lifecycle
'pack' event, which generates a release tarball. This tarball is
then extracted, and the contents of the configured directory
are then uploaded to docs-draft. A success pattern has been
included, so a passing build will link to the newly constructed
application directly.

Change-Id: I24c55738025284d9e0baef13ec8a9b49a90d4ef7